VirtualScroller: Rename `virtual-scroller.js` to `VirtualScroller.js`.
