+++ smoke/flat/script.js
@@ -1,7 +1,7 @@
+import {VirtualScroller} from '../../src/VirtualScroller.js';
 import {Layout1d} from '../../src/layouts/Layout1d.js';
 import {scroller} from '../../lit-html/lit-scroller.js';
 import {html, render} from '../../node_modules/lit-html/lib/lit-extended.js';
-import {VirtualScroller} from '../../src/virtual-scroller.js';
 
 const items = new Array(200).fill({name: 'item'});
 const container = document.getElementById('container');
