+++ src/virtual-scroller-element.js
@@ -1,7 +1,7 @@
 import {ItemSource, _item, _key} from './ItemSource.js';
 import {Layout1dGrid} from './layouts/Layout1dGrid.js';
 import {Layout1d} from './layouts/Layout1d.js';
-import {VirtualScroller} from './virtual-scroller.js';
+import {VirtualScroller} from './VirtualScroller.js';
 
 /** Properties */
 const _scroller = Symbol();
